update vue tests
