+++ libs/vue/src/generators/component/files/__name__.vue.template
@@ -6,7 +6,7 @@
 import { defineComponent } from 'vue';
 
 export default defineComponent({
-  name: '<%= name %>',
+  name: '<%= fileName %>',
 });
 </script>
