+++ libs/vue/src/generators/vuex/schema.json
@@ -1,5 +1,6 @@
 {
   "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
   "$id": "Vuex",
   "title": "Vuex Options Schema",
   "type": "object",
