+++ libs/vue/src/index.ts
@@ -1,4 +1,4 @@
 export { applicationGenerator } from './generators/application/generator';
-export { libraryGenerator } from './generators/library/generator';
 export { componentGenerator } from './generators/component/generator';
+export { libraryGenerator } from './generators/library/generator';
 export { vuexGenerator } from './generators/vuex/generator';
